+++ scripts/read_distribution.py
@@ -14,8 +14,7 @@ from typing import NamedTuple
 from bx.intervals import Intersecter, Interval
 
 from rseqc import BED, SAM, bam_cigar
-from rseqc.cli_common import add_input_bam_arg, add_refgene_arg, create_parser, validate_files_exist
-from rseqc.SAM import _pysam_iter
+from rseqc.cli_common import _pysam_iter, add_input_bam_arg, add_refgene_arg, create_parser, validate_files_exist
 
 
 def cal_size(list: list[list]) -> int: